+++ src/components/geojson-input.vue
@@ -1,231 +1,44 @@
 <template>
-  <div>
-    <table>
-      <thead>
-        <tr>
-          <td class="text-bold">
-            {{paramName || '?'}}
-          </td>
-          <td class="text-weight-thin" style="text-align:right">
-            {{paramType}}
-          </td>
-        </tr>
-      </thead>
-      <tbody>
-      <tr>
-        <td class="gjMap">
-          <div
-            class="absolute-top-left map-buttons column items-start"
-          >
-            <div
-              v-if="!$q.platform.is.mobile"
-              class="column q-mb-md"
-            >
-              <q-btn
-                dense glossy color="indigo-11"
-                icon="add"
-                class="shadow-8"
-                @click="doZoom(false)"
-              >
-                <q-tooltip anchor="center left" self="center right" :delay="1000">
-                  Zoom in
-                </q-tooltip>
-              </q-btn>
-              <q-btn
-                v-if="!$q.platform.is.mobile"
-                dense glossy color="indigo-11"
-                icon="remove"
-                class="shadow-8"
-                @click="doZoom(true)"
-              >
-                <q-tooltip anchor="center left" self="center right" :delay="1000">
-                  Zoom out
-                </q-tooltip>
-              </q-btn>
-            </div>
-
-            <q-btn
-              dense glossy color="indigo-11"
-              icon="center_focus_weak"
-              class="shadow-8"
-              @click="zoomToAll()"
-            >
-              <q-tooltip anchor="center left" self="center right" :delay="1000">
-                Zoom to all apositions
-              </q-tooltip>
-            </q-btn>
-          </div>
-
-          <l-map
-            ref="gjMap"
-            style="height: 100%; width: 100%"
-            :zoom="zoom"
-            :center="center"
-            :options="{zoomControl:false}"
-          >
-<!--            <l-control-polyline-measure :options="{ showUnitControl: true }" position="bottomright"/>-->
-
-            <l-feature-group
-              ref="featureGroup"
-            >
-              <l-circle-marker
-                v-if="paramType === 'Point' && point.length"
-                :lat-lng="point"
-              />
-
-              <l-circle-marker
-                v-if="paramType === 'MultiPoint' && points.length"
-                v-for="(p, index) in points" :key="`${p[0]},${p[1]},${index}`"
-                :lat-lng="p"
-              />
-
-              <l-polygon
-                v-if="paramType === 'Polygon' && polygon.length"
-                :lat-lngs="polygon"
-              />
-
-            </l-feature-group>
-
-            <div v-if="mousePos">
-              <l-circle-marker
-                :lat-lng="mousePos.latLon"
-                color="yellow"
-                :radius="mousePos.radius"
-                dash-array="5 4"
-              />
-              <l-circle-marker
-                :lat-lng="mousePos.latLon"
-                :color="mousePos.color"
-                :radius="mousePos.radius * 3"
-                dash-array="14 6"
-              />
-            </div>
-          </l-map>
-        </td>
-
-        <td style="vertical-align:top">
-          <q-scroll-area
-            style="width:200px; height: 400px;"
-            :thumb-style="{ background: 'blue', borderRadius: '5px' }"
-          >
-            <position-table
-              v-if="paramType === 'Point' && point.length"
-              :lat-lons="[point]"
-            />
-
-            <position-table
-              v-else-if="paramType === 'MultiPoint' && points.length"
-              :lat-lons="points"
-            />
-
-            <position-table
-              v-else-if="paramType === 'Polygon' && polygon.length"
-              :lat-lons="polygon"
-            />
-
-            <position-table
-              v-else
-              :lat-lons="[]"
-            />
-
-            <div style="font-size:x-small;color:gray">
-              <pre v-if="point.length">point={{point}}</pre>
-              <pre v-if="points.length">points={{points}}</pre>
-              <pre v-if="polygon.length">polygon={{polygon}}</pre>
-            </div>
-          </q-scroll-area>
-        </td>
-      </tr>
-      </tbody>
-      <tfoot v-if="!readonly">
-      <tr>
-        <td colspan="2" style="text-align:right">
-          <q-btn
-            v-if="defaultValue"
-            icon="settings_backup_restore"
-            dense round
-            @click="setFeatureDataAndEmit(defaultValue)"
-          >
-            <q-tooltip>Reset to default value</q-tooltip>
-          </q-btn>
-          <q-btn
-            v-if="defaultValue !== undefined"
-            icon="clear"
-            dense round
-            @click="setFeatureDataAndEmit()"
-          >
-            <q-tooltip>Clear all</q-tooltip>
-          </q-btn>
-        </td>
-      </tr>
-      </tfoot>
-    </table>
-    <span style="font-family:monospace;font-size:x-small;color:gray">valueString='{{valueString}}'</span>
+  <div style="max-width:400px">
+    <pre>entry_id = {{entry_id}}</pre>
+    <qgeomap
+      ref="qgeomap"
+      :editable="!readonly"
+      v-on:warning="showWarning"
+      include-table
+      style="height:600px; width:400px"
+    />
+    <span style="word-break: break-all">valueString = {{valueString}}</span>
   </div>
 </template>
 
 <script>
-  import Vue from 'vue'
-  import L from 'leaflet'
+  import map from 'lodash/map'
 
-  import {
-    LMap,
-    LMarker,
-    LFeatureGroup,
-    LLayerGroup,
-    LPolyline,
-    LPolygon,
-    LPopup,
-    LTooltip,
-    LCircle,
-    LCircleMarker,
-  } from 'vue2-leaflet'
-
-  // import 'leaflet-draw'
-  // import LControlPolylineMeasure from 'vue2-leaflet-polyline-measure'
-
-  import {} from 'vue2-leaflet-googlemutant'
-
-  // require('leaflet-draw/dist/leaflet.draw')
-  import _ from 'lodash'
-
-  import PositionTable from 'components/position-table'
-
-  const debug = false
+  const debug = true
 
   export default {
-    components: {
-      LMap,
-      LMarker,
-      LFeatureGroup,
-      LLayerGroup,
-      LPolyline,
-      LPolygon,
-      LPopup,
-      LTooltip,
-      LCircle,
-      LCircleMarker,
-      // LControlPolylineMeasure,
-      PositionTable
-    },
-
     props: {
       paramName: {
         type: String,
-        required: false
-      },
-      paramType: {
-        type: String,
         required: true
       },
+
+      paramType: {
+        type: String,
+        default: 'Point'
+      },
+
       value: {
         type: String,
         required: false
       },
+
       defaultValue: {
         type: String,
         required: false
       },
+
       readonly: {
         type: Boolean,
         required: false,
@@ -233,46 +46,36 @@
       },
     },
 
+    computed: {
+      entry_id() {
+        return `${this.paramName}:${this.paramType}`
+      }
+    },
+
     data () {
       return {
-        center: [36.83, -121.9],
-        zoom: 10,
-        mousePos: null,
-
-        point: [],
-        points: [],
-        polygon: [],
-
         valueString: ''
       }
     },
 
-    created () {
-      this.$root.$on('evt-map-center-at', this.centerMapAt)
-      this.$root.$on('evt-map-on-mouse-pos', this.onMousePos)
-    },
-
-    destroyed () {
-      this.$root.$off('evt-map-center-at', this.centerMapAt)
-      this.$root.$off('evt-map-on-mouse-pos', this.onMousePos)
-    },
-
     mounted() {
+      if (debug) console.log(`MOUNTED geojson-input
+      paramName=${this.paramName}
+      paramType=${this.paramType}
+      value=${this.value}
+      defaultValue=${this.defaultValue}
+      readonly=${this.readonly}
+      `)
       this.setFeatureData(this.value)
-      this.$nextTick(this.initMap)
     },
 
     methods: {
-      setFeatureDataAndEmit(value) {
-        const json = this.setFeatureData(value)
-        this.updateValueString(json)
-      },
-
       setFeatureData(value) {
+        const qgeomap = this.$refs.qgeomap
+
         this.valueString = value || ''
-        this.point.splice(0)
-        this.points.splice(0)
-        this.polygon.splice(0)
+
+        const entry_id = this.entry_id
 
         let json
         if (this.valueString.trim()) {
@@ -280,20 +83,45 @@
             json = JSON.parse(this.valueString)
             switch (this.paramType) {
               case 'Point': {
-                this.point = json
-                if (debug) console.debug(`setFeatureData: paramType=${this.paramType} point=`, this.point)
-                break
-              }
-
-              case 'MultiPoint': {
-                this.points = json
-                if (debug) console.debug(`setFeatureData: paramType=${this.paramType} points=`, this.points)
+                if (debug) console.debug(`setFeatureData: paramType=${this.paramType} point=`, json)
+                const coordinates = [json[1], json[0]]
+                qgeomap.addEntry({
+                  entry_id,
+                  geometry: {
+                    type: "Feature",
+                    geometry: {
+                      type: "Point",
+                      coordinates,
+                    },
+                  },
+                  color: 'cyan',
+                  tooltip: entry_id,
+                })
                 break
               }
 
               case 'Polygon': {
-                this.polygon = json
-                if (debug) console.debug(`setFeatureData: paramType=${this.paramType } polygon=`, this.polygon)
+                if (debug) console.debug(`setFeatureData: paramType=${this.paramType } polygon=`, json)
+
+                const coordinates = [ map(json, ([lat, lon]) => [lon, lat]) ]
+
+                qgeomap.addEntry({
+                  entry_id,
+                  geometry: {
+                    type: "Feature",
+                    geometry: {
+                      type: "Polygon",
+                      coordinates,
+                    },
+                  },
+                  color: 'yellow',
+                  tooltip: 'Polygon',
+                })
+                break
+              }
+
+              case 'MultiPoint': {
+                if (debug) console.debug(`setFeatureData: paramType=${this.paramType} json=`, json)
                 break
               }
 
@@ -307,355 +135,21 @@
         return json
       },
 
-      initMap() {
-        const map = this.$refs.gjMap.mapObject
-
-        if (debug) console.debug(`geojson-input initMap: paramName=${this.paramName}`, 'map=', map)
-
-        this.$mapMan(map)
-
-        if (!this.readonly) {
-          this.initEditor()
-        }
-      },
-
-      initEditor() {
-        const map = this.$refs.gjMap.mapObject
-        const featureGroup = this.$refs.featureGroup.mapObject
-        if (debug) console.debug('initEditor: featureGroup=', featureGroup)
-
-        let circle = false
-        let circlemarker = false  // {icon: new MyCustomMarker()}
-        let marker = false  // {icon: new MyCustomMarker()}
-        let rectangle = false
-        let polyline = false
-        let polygon = false
-
-        let repeatMode = false
-
-        const enableMarker = () => {
-          marker = {
-            repeatMode,
-          }
-        }
-
-        const enableCircleMarker = () => {
-          circlemarker = {
-            shapeOptions: {
-              weight: 4
-            },
-            repeatMode,
-          }
-        }
-
-        const enablePolyline = () => {
-          polyline = {
-            shapeOptions: {
-              color: '#f357a1',
-              weight: 4
-            }
-          }
-        }
-
-        const enableRectangle = () => {
-          rectangle = {
-            shapeOptions: {
-              color: '#f357a1',
-              weight: 4
-            }
-          }
-        }
-
-        const enablePolygon = () => {
-          polygon = {
-            allowIntersection: false, // Restricts shapes to simple polygons
-            drawError: {
-              color: '#e1e100', // Color the shape will turn when intersects
-            },
-            shapeOptions: {
-              color: '#bada55'
-            }
-          }
-        }
-
-        switch (this.paramType) {
-          case 'Point':
-            repeatMode = true
-            enableMarker()
-            enableCircleMarker()
-            break
-
-          case 'MultiPoint':
-            repeatMode = true
-            enableMarker()
-            enableCircleMarker()
-            break
-
-          case 'LineString':
-            enablePolyline()
-            break
-
-          case 'MultiLineString':
-            enablePolyline()
-            break
-
-          case 'Polygon':
-            enableRectangle()
-            enablePolygon()
-            break
-
-          case 'MultiPolygon':
-            repeatMode = true
-            enableRectangle()
-            enablePolygon()
-            break
-
-          case 'GeometryCollection':
-            repeatMode = true
-            enableMarker()
-            enableCircleMarker()
-            enablePolyline()
-            enableRectangle()
-            enablePolygon()
-            break
-
-          case 'Feature':
-          case 'FeatureCollection':
-            break
-
-          case 'GeoJSON':
-            break
-        }
-
-        const options = {
-          position: 'topright',
-          draw: {
-            circle,
-            circlemarker,
-            marker,
-            rectangle,
-            polyline,
-            polygon,
-          },
-          edit: {
-            featureGroup,
-            edit: {
-              selectedPathOptions: {
-                maintainColor: true,
-                opacity: 0.5
-              }
-            },
-            remove: true,
-          }
-        }
-
-        const drawControl = new L.Control.Draw(options)
-        map.addControl(drawControl)
-
-        map.on(L.Draw.Event.CREATED, e => {
-          this.layerCreated(e.layer)
-        })
-
-        map.on(L.Draw.Event.EDITED, e => {
-          this.layersEdited()
-        })
-
-        map.on(L.Draw.Event.DELETED, e => {
-          this.layersDeleted()
-        })
-      },
-
-      layerCreated(layer) {
-        console.debug('layerCreated:', 'layer=', layer)
-
-        let json = null
-        switch (this.paramType) {
-          case 'Point': {
-            if (layer._latlng) {
-              this.point = [layer._latlng.lat, layer._latlng.lng]
-              console.debug('layerCreated: point=', this.point)
-              json = this.point
-            }
-            break
-          }
-
-          case 'MultiPoint': {
-            if (layer._latlng) {
-              this.points.push([layer._latlng.lat, layer._latlng.lng])
-              json = this.points
-              console.debug('layerCreated: points=', this.points)
-            }
-            break
-          }
-
-          case 'Polygon': {
-            if (layer._latlngs && layer._latlngs.length) {
-              this.polygon = _.map(layer._latlngs[0], ({lat, lng}) => [lat, lng])
-              console.debug('layerCreated: polygon=', this.polygon)
-              json = this.polygon
-            }
-            break
-          }
-
-          // TODO the other paramType's
-        }
-
-        this.updateValueString(json)
-      },
-
-      layersEdited() {
-        const featureGroup = this.$refs.featureGroup.mapObject
-        let json = null
-        switch (this.paramType) {
-          case 'Point': {
-            this.point.splice(0)
-            const layer = _.head(featureGroup.getLayers())
-            if (layer && layer._latlng) {
-              this.point = [layer._latlng.lat, layer._latlng.lng]
-            }
-            json = this.point
-            break
-          }
-
-          case 'MultiPoint': {
-            this.points.splice(0)
-            featureGroup.eachLayer(layer => {
-              if (layer && layer._latlng) {
-                this.points.push([layer._latlng.lat, layer._latlng.lng])
-              }
-            })
-            console.debug('layersEdited: points=', this.points)
-            json = this.points
-            break
-          }
-
-          case 'Polygon': {
-            this.polygon.splice(0)
-            const layer = _.head(featureGroup.getLayers())
-            if (layer && layer._latlngs && layer._latlngs.length) {
-              _.each(layer._latlngs[0], ({lat, lng}) => {
-                this.polygon.push([lat, lng])
-              })
-            }
-            json = this.polygon
-            break
-          }
-
-          // TODO the other paramType's
-        }
-
-        this.updateValueString(json)
-      },
-
-      layersDeleted() {
-        const featureGroup = this.$refs.featureGroup.mapObject
-        let json = null
-        switch (this.paramType) {
-          case 'Point':
-            this.point.splice(0)
-            console.debug('layersDeleted: point=', this.point)
-            json = this.point
-            break
-
-          case 'MultiPoint':
-            // just grab all remaining points in featureGroup:
-            this.points.splice(0)
-            featureGroup.eachLayer(layer => {
-              console.debug('::: featureGroup layer._latlng=', layer._latlng)
-              if (layer._latlng) {
-                this.points.push([layer._latlng.lat, layer._latlng.lng])
-              }
-            })
-            console.debug('layersDeleted: points=', this.points)
-            json = this.points
-            break
-
-          case 'Polygon':
-            this.polygon.splice(0)
-            console.debug('layersDeleted: polygon=', this.polygon)
-            json = this.polygon
-            break
-
-          // TODO the other paramType's
-        }
-
-        this.updateValueString(json)
-      },
-
       updateValueString(json) {
         this.valueString = json && stringify(json) || ''
         console.debug(`::: updateValueString emiting '${this.valueString}'`)
         this.$emit('input', this.valueString)
       },
 
-      centerMapAt (latLon) {
-        this.center = latLon
-      },
-
-      onMousePos (latLon) {
-        if (latLon) {
-          this.mousePos = {
-            latLon,
-            color: '#ff0000',
-            radius: 5,
-          }
-        }
-        else {
-          this.mousePos = null
-        }
-      },
-
-      doZoom (out) {
-        const map = this.$refs.gjMap.mapObject
-        out ? map.zoomOut() : map.zoomIn()
-      },
-
-      zoomToAll () {
-        const map = this.$refs.gjMap.mapObject
-
-        let thePoints = []
-        switch (this.paramType) {
-          case 'Point': {
-            if (this.point.length) {
-              thePoints = [this.point]
-            }
-            break
-          }
-
-          case 'MultiPoint': {
-            if (this.points.length) {
-              thePoints = this.points
-            }
-            break
-          }
-
-          case 'Polygon': {
-            if (this.polygon.length) {
-              thePoints = this.polygon
-            }
-            break
-          }
-
-          // TODO the other paramType's
-        }
-
-        if (thePoints.length > 1) {
-          const min = _.reduce(thePoints, (min, [lat, lon]) => ({
-            lat: min.lat != null ? Math.min(min.lat, lat) : lat,
-            lon: min.lon != null ? Math.min(min.lon, lon) : lon,
-          }), {lat:null, lon:null})
-          const max = _.reduce(thePoints, (max, [lat, lon]) => ({
-            lat: max.lat != null ? Math.max(max.lat, lat) : lat,
-            lon: max.lon != null ? Math.max(max.lon, lon) : lon,
-          }), {lat:null, lon:null})
-          if (debug) console.debug('zoomToAll: min=', min, 'max=', max)
-          const bounds = [[max.lat, max.lon], [min.lat, min.lon]]
-          map.fitBounds(bounds, { padding: [20, 20], maxZoom: 13, animate:true })
-        }
-        else if (thePoints.length === 1) {
-          map.setView(thePoints[0], 11, {animate:true})
-          this.centerMapAt(thePoints[0])
-        }
+      showWarning(message) {
+        console.warn('WARN:', message)
+        this.$q.notify({
+          message,
+          position: 'top',
+          color: 'warning',
+          textColor: 'black',
+          timeout: 1500,
+        })
       },
     },
   }
@@ -674,42 +168,3 @@
   }
 
 </script>
-
-<style src="leaflet/dist/leaflet.css" />
-<!--<style src="leaflet-measure/dist/leaflet-measure.css" />-->
-<style src="leaflet-draw/dist/leaflet.draw.css" />
-
-<style>
-  .gjMap {
-    width: 500px;
-    height: 400px;
-    /*border: 1px solid red;*/
-  }
-
-  .leaflet-container.my-default-cursor {
-    cursor: default;
-  }
-
-  .leaflet-control-mouseposition {
-    font-family: monospace, serif !important;
-    font-size: smaller !important;
-    background-color: rgba(255, 255, 255, 0.75) !important;
-    padding: 0 4px 0 4px !important;
-    border: 1px solid lightgray;
-    border-radius: 4px;
-    box-shadow: 0 5px 5px -3px rgba(0,0,0,0.2), 0 8px 10px 1px rgba(0,0,0,0.14), 0 3px 14px 2px rgba(0,0,0,0.12);
-  }
-
-  .leaflet-control-measure h3, .leaflet-measure-resultpopup h3 {
-    font-size:1em !important;
-    font-weight: bold;
-  }
-</style>
-
-<style scoped>
-  .map-buttons {
-    z-index: 9999 !important;
-    margin-left: 25px;
-    margin-top: 55px;
-  }
-</style>